Better modularize the code
