+++ server/.jshintrc
@@ -15,10 +15,15 @@
     "Npm": false,
     "Package": false,
 
-
     // Our dependencies.
     "_": false,
     "Mongo": false,
-    "WebApp": false
+    "WebApp": false,
+
+    // Global functions.
+    "createBinaries": true,
+    "launchApp": true,
+    "platformSpecificSetting": true,
+    "serveBuild": true
   }
 }
