+++ server/platformSpecificSetting.js
@@ -1,0 +1,6 @@
+platformSpecificSetting = function(settings) {
+  switch (process.platform) {
+    case 'darwin':
+      return settings.osx;
+  }
+};
